requirements.txt
numpy
torch
torch_geometric
egg  # install via git
pyyaml
networkx
matplotlib
pandas
python-Levenshtein
python-louvain
